Bug #3985 Fix NPE in External Storage Handler
Problem:
When a query type wasn’t selected and the link “Add attribute” was
clicked, a Null Pointer Exception happened.
It happened because the link creation has an “onLinkActivated” method
which requires “queryType” as a parameter.

Solution:
I added not null check before the link creation. Now the link appears
only after the query type is selected in the drop-down list.